+++ app/controllers/server.rb
@@ -1,8 +1,8 @@
-require_relative '../models/check_client_params'
+require_relative '../models/params_checker'
 
 module RushHour
   class Server < Sinatra::Base
-    include CheckClientParams
+    include ParamsChecker
 
     post '/sources' do
       client = Client.new(params[:client])
