+++ test/test_helper.rb
@@ -29,9 +29,9 @@ module TestHelpers
   end
 
   def setup_data
-    [PayloadRequest.create(url: Url.create(address: "http://jumpstartlab.com"),
+    [PayloadRequest.create(url: Url.find_or_create_by(address: "http://jumpstartlab.com"),
                                referrer: Referrer.create(address: "http://amazon.com"),
-                               request_type: RequestType.create(verb: "GET"),
+                               request_type: RequestType.find_or_create_by(verb: "GET"),
                                event: Event.create(name: "facebook"),
                                user_agent: UserAgent.create(browser: "Mozilla", platform: "Windows"),
                                resolution: Resolution.create(width: "2560", height: "1440"),
@@ -39,9 +39,9 @@ module TestHelpers
                                requested_at: "2013-02-16 21:40:00 -0700",
                                responded_in: 20
                               ),
-    PayloadRequest.create(url: Url.create(address: "http://turing.io"),
+    PayloadRequest.create(url: Url.find_or_create_by(address: "http://turing.io"),
                                referrer: Referrer.create(address: "http://amazon.com"),
-                               request_type: RequestType.create(verb: "GET"),
+                               request_type: RequestType.find_or_create_by(verb: "GET"),
                                event: Event.create(name: "facebook"),
                                user_agent: UserAgent.create(browser: "Chrome", platform: "Webkit"),
                                resolution: Resolution.create(width: "1920", height: "1280"),
@@ -49,9 +49,9 @@ module TestHelpers
                                requested_at: "2013-02-16 21:37:00 -0700",
                                responded_in: 30
                               ),
-    PayloadRequest.create(url: Url.create(address: "http://jumpstartlab.com"),
+    PayloadRequest.create(url: Url.find_or_create_by(address: "http://jumpstartlab.com"),
                                referrer: Referrer.create(address: "http://newegg.com"),
-                               request_type: RequestType.create(verb: "POST"),
+                               request_type: RequestType.find_or_create_by(verb: "POST"),
                                event: Event.create(name: "twitter"),
                                user_agent: UserAgent.create(browser: "Safari", platform: "Macintosh"),
                                resolution: Resolution.create(width: "1920", height: "1280"),
@@ -59,9 +59,9 @@ module TestHelpers
                                requested_at: "2013-02-16 21:38:00 -0700",
                                responded_in: 40
                               )
-    # PayloadRequest.create(url: Url.create(address: "http://today.turing.io"),
+    # PayloadRequest.create(url: Url.find_or_create_by(address: "http://today.turing.io"),
     #                            referrer: Referrer.create(address: "http://jumpstartlab.com"),
-    #                            request_type: RequestType.create(verb: "GET"),
+    #                            request_type: RequestType.find_or_create_by(verb: "GET"),
     #                            event: Event.create(name: "socialLogin"),
     #                            user_agent: UserAgent.create(browser: "Chrome", platform: "Macintosh"),
     #                            resolution: Resolution.create(width: "1920", height: "1280"),
@@ -83,9 +83,9 @@ module TestHelpers
   end
 
   def referrer_data #REDO
-    [PayloadRequest.create(url: Url.create(address: "http://jumpstartlab.com"),
+    [PayloadRequest.create(url: Url.find_or_create_by(address: "http://jumpstartlab.com"),
                                referrer: Referrer.create(address: "http://amazon.com"),
-                               request_type: RequestType.create(verb: "GET"),
+                               request_type: RequestType.find_or_create_by(verb: "GET"),
                                event: Event.create(name: "facebook"),
                                user_agent: UserAgent.create(browser: "Mozilla", platform: "Windows"),
                                resolution: Resolution.create(width: "2560", height: "1440"),
@@ -93,9 +93,9 @@ module TestHelpers
                                requested_at: "2013-02-16 21:40:00 -0700",
                                responded_in: 20
                               ),
-    PayloadRequest.create(url: Url.create(address: "http://turing.io"),
+    PayloadRequest.create(url: Url.find_or_create_by(address: "http://turing.io"),
                                referrer: Referrer.create(address: "http://amazon.com"),
-                               request_type: RequestType.create(verb: "GET"),
+                               request_type: RequestType.find_or_create_by(verb: "GET"),
                                event: Event.create(name: "facebook"),
                                user_agent: UserAgent.create(browser: "Chrome", platform: "Webkit"),
                                resolution: Resolution.create(width: "1920", height: "1280"),
@@ -103,9 +103,9 @@ module TestHelpers
                                requested_at: "2013-02-16 21:37:00 -0700",
                                responded_in: 30
                               ),
-    PayloadRequest.create(url: Url.create(address: "http://jumpstartlab.com"),
+    PayloadRequest.create(url: Url.find_or_create_by(address: "http://jumpstartlab.com"),
                                referrer: Referrer.create(address: "http://newegg.com"),
-                               request_type: RequestType.create(verb: "POST"),
+                               request_type: RequestType.find_or_create_by(verb: "POST"),
                                event: Event.create(name: "twitter"),
                                user_agent: UserAgent.create(browser: "Safari", platform: "Macintosh"),
                                resolution: Resolution.create(width: "1920", height: "1280"),
@@ -113,9 +113,9 @@ module TestHelpers
                                requested_at: "2013-02-16 21:38:00 -0700",
                                responded_in: 40
                               ),
-    PayloadRequest.create(url: Url.create(address: "http://turing.io"),
+    PayloadRequest.create(url: Url.find_or_create_by(address: "http://turing.io"),
                                referrer: Referrer.create(address: "http://jumpstartlab.com"),
-                               request_type: RequestType.create(verb: "GET"),
+                               request_type: RequestType.find_or_create_by(verb: "GET"),
                                event: Event.create(name: "socialLogin"),
                                user_agent: UserAgent.create(browser: "Chrome", platform: "Macintosh"),
                                resolution: Resolution.create(width: "1920", height: "1280"),
@@ -123,9 +123,9 @@ module TestHelpers
                                requested_at: "2013-02-15 21:38:28 -0700",
                                responded_in: 10
                               ),
-    PayloadRequest.create(url: Url.create(address: "http://turing.io"),
+    PayloadRequest.create(url: Url.find_or_create_by(address: "http://turing.io"),
                                referrer: Referrer.create(address: "http://jumpstartlab.com"),
-                               request_type: RequestType.create(verb: "GET"),
+                               request_type: RequestType.find_or_create_by(verb: "GET"),
                                event: Event.create(name: "socialLogin"),
                                user_agent: UserAgent.create(browser: "Chrome", platform: "Macintosh"),
                                resolution: Resolution.create(width: "1920", height: "1280"),
@@ -134,16 +134,16 @@ module TestHelpers
                                responded_in: 10
                               )]
 
-    # PayloadRequest.create(url: Url.create(address: "http://turing.io"), referrer: Referrer.create(address:"http://newegg.com"))
-    # PayloadRequest.create(url: Url.create(address: "http://turing.io"), referrer: Referrer.create(address:"http://newegg.com"))
-    # PayloadRequest.create(url: Url.create(address: "http://turing.io"), referrer: Referrer.create(address:"http://newegg.com"))
-    # PayloadRequest.create(url: Url.create(address: "http://turing.io"), referrer: Referrer.create(address:"http://newegg.com"))
-    # PayloadRequest.create(url: Url.create(address: "http://turing.io"), referrer: Referrer.create(address:"http://amazon.com"))
-    # PayloadRequest.create(url: Url.create(address: "http://turing.io"), referrer: Referrer.create(address:"http://amazon.com"))
-    # PayloadRequest.create(url: Url.create(address: "http://turing.io"), referrer: Referrer.create(address:"http://amazon.com"))
-    # PayloadRequest.create(url: Url.create(address: "http://turing.io"), referrer: Referrer.create(address:"http://nba.com"))
-    # PayloadRequest.create(url: Url.create(address: "http://turing.io"), referrer: Referrer.create(address:"http://nba.com"))
-    # PayloadRequest.create(url: Url.create(address: "http://turing.io"), referrer: Referrer.create(address:"http://nfl.com"))
+    # PayloadRequest.create(url: Url.find_or_create_by(address: "http://turing.io"), referrer: Referrer.create(address:"http://newegg.com"))
+    # PayloadRequest.create(url: Url.find_or_create_by(address: "http://turing.io"), referrer: Referrer.create(address:"http://newegg.com"))
+    # PayloadRequest.create(url: Url.find_or_create_by(address: "http://turing.io"), referrer: Referrer.create(address:"http://newegg.com"))
+    # PayloadRequest.create(url: Url.find_or_create_by(address: "http://turing.io"), referrer: Referrer.create(address:"http://newegg.com"))
+    # PayloadRequest.create(url: Url.find_or_create_by(address: "http://turing.io"), referrer: Referrer.create(address:"http://amazon.com"))
+    # PayloadRequest.create(url: Url.find_or_create_by(address: "http://turing.io"), referrer: Referrer.create(address:"http://amazon.com"))
+    # PayloadRequest.create(url: Url.find_or_create_by(address: "http://turing.io"), referrer: Referrer.create(address:"http://amazon.com"))
+    # PayloadRequest.create(url: Url.find_or_create_by(address: "http://turing.io"), referrer: Referrer.create(address:"http://nba.com"))
+    # PayloadRequest.create(url: Url.find_or_create_by(address: "http://turing.io"), referrer: Referrer.create(address:"http://nba.com"))
+    # PayloadRequest.create(url: Url.find_or_create_by(address: "http://turing.io"), referrer: Referrer.create(address:"http://nfl.com"))
   end
 end
